+++ src/client/credential-store.ts
@@ -1,7 +1,7 @@
 import { mkdir, readFile, rename, unlink, writeFile } from "node:fs/promises";
 import { dirname, join } from "node:path";
 import { homedir } from "node:os";
-import { createWindowsDpapiCodec } from "../platform/windows-dpapi";
+import { createWindowsDpapiCodec } from "./windows-dpapi";
 
 interface StoredCredential {
   v: 1;