+++ src/linux-node/index.ts
@@ -1,3 +1,0 @@
-export * from "./config";
-export * from "./credential-store";
-export * from "./runtime";